# Conflicts:
#	out/production/Lemonade Stand/com/imerir/Game.class
#	out/production/Lemonade Stand/com/imerir/MainWindow.class
#	out/production/Lemonade Stand/com/imerir/Player.class
#	out/production/Lemonade Stand/com/imerir/SellFactory$1.class
#	out/production/Lemonade Stand/com/imerir/SellFactory.class

+++ src/com/imerir/Player.java
@@ -1,14 +1,14 @@
 package com.imerir;
 
 public class Player {
-    private int money;
+    private double money;
     public Player(int money){
         this.money = money;
     }
     public void setMoney(double money){
         this.money += money;
     }
-    public int getMoney() {
+    public double getMoney() {
         return money;
     }
 }
